Talent management conversion
Converted to functional components
