+++ services/frontend-v3/src/components/talentManagement/TalentManagementView.jsx
@@ -3,8 +3,8 @@ import { FormattedMessage, injectIntl } from "react-intl";
 
 import { Row, Col, List } from "antd";
 
-class TalentManagementView extends Component {
-  generateTalentManagementInfoList(dataSource) {
+function TalentManagementView(props) {
+  const generateTalentManagementInfoList = dataSource => {
     return (
       <List
         itemLayout="horizontal"
@@ -16,41 +16,17 @@ class TalentManagementView extends Component {
         )}
       />
     );
-  }
+  };
 
-  getTalentManagementInfo(locale) {
-    const data = this.props.data;
+  const info = props.info;
 
-    const careerMobility = {
-      title: <FormattedMessage id="profile.career.mobility" />,
-      description: data.careerMobility.description[locale] || (
-        <FormattedMessage id="profile.not.specified" />
-      )
-    };
-
-    const talentMatrixResult = {
-      title: <FormattedMessage id="profile.talent.matrix.result" />,
-      description: data.talentMatrixResult.description[locale] || (
-        <FormattedMessage id="profile.not.specified" />
-      )
-    };
-
-    return [careerMobility, talentMatrixResult];
-  }
-
-  render() {
-    const locale = this.props.intl.formatMessage({ id: "language.code" });
-
-    const info = this.getTalentManagementInfo(locale);
-
-    return (
-      <Row>
-        <Col xs={24} lg={24}>
-          {this.generateTalentManagementInfoList(info)}
-        </Col>
-      </Row>
-    );
-  }
+  return (
+    <Row>
+      <Col xs={24} lg={24}>
+        {generateTalentManagementInfoList(info)}
+      </Col>
+    </Row>
+  );
 }
 
 export default injectIntl(TalentManagementView);
